requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
fastapi==0.112.2
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.8
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
oauthlib==3.2.2
packaging==23.2
pefile==2024.8.26
persiantools==4.2.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.7.1
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
websockets==13.0
win32printing==0.1.3
